+++ src/utils/__mocks__/quasar-lang.ts
@@ -1,6 +1,0 @@
-export const quasarLangList = {
-  // eslint-disable-next-line no-undef
-  '../../node_modules/quasar/lang/en-US.mjs': () => new Promise((resolve) => { resolve({ default: 'en' }); }),
-  // eslint-disable-next-line no-undef
-  '../../node_modules/quasar/lang/fr.mjs': () => new Promise((resolve) => { resolve({ default: 'fr' }); }),
-};