+++ test/layouts/AppLayout.spec.ts
@@ -6,31 +6,43 @@ import { config, mount, shallowMount } from '@vue/test-utils';
 import AppLayout from 'layouts/AppLayout.vue';
 import { QBtn, QToolbarTitle } from 'quasar';
 import { i18n } from 'boot/i18n';
+import { createTestingPinia } from '@pinia/testing';
 
 installQuasarPlugin();
 
 jest.mock('boot/i18n');
-jest.mock('src/utils/quasar-lang');
 
 describe('layouts/AppLayout', () => {
   config.global.mocks.$t = i18n.global.t;
   config.global.plugins = [...config.global.plugins, i18n];
 
   it('sets the correct default data', () => {
-    const { vm } = shallowMount(AppLayout);
+    const { vm } = shallowMount(AppLayout, {
+      global: {
+        plugins: [createTestingPinia()],
+      },
+    });
 
     expect(typeof vm.leftDrawerOpen).toBe('boolean');
     expect(vm.leftDrawerOpen).toBeFalsy();
   });
 
   it('has toggleLeftDrawer method', () => {
-    const { vm } = shallowMount(AppLayout);
+    const { vm } = shallowMount(AppLayout, {
+      global: {
+        plugins: [createTestingPinia()],
+      },
+    });
 
     expect(typeof vm.toggleLeftDrawer).toBe('function');
   });
 
   it('correctly toggle LeftDrawer when menu button is pressed', async () => {
-    const wrapper = mount(AppLayout);
+    const wrapper = mount(AppLayout, {
+      global: {
+        plugins: [createTestingPinia()],
+      },
+    });
     const { vm } = wrapper;
 
     expect(vm.leftDrawerOpen).toBeFalsy();
@@ -42,7 +54,11 @@ describe('layouts/AppLayout', () => {
   });
 
   it('display application name in header', async () => {
-    const wrapper = mount(AppLayout);
+    const wrapper = mount(AppLayout, {
+      global: {
+        plugins: [createTestingPinia()],
+      },
+    });
 
     const toolbarTitle = wrapper.findComponent(QToolbarTitle);
     expect(toolbarTitle.vm.$el.textContent).toContain('MHRS - Gestionnaire de talismans');